rust-passes: widen mixed-minimal-width operands at binary-op boundaries
The typer tracks value ranges, not declared widths: `q * 4` on a
`Uint<32>` types as `Uint<0..17179869181>` and the rust backend
deliberately renders it at u64. When such a range-widened expression
met a narrower one inside a comparison, the typer had already wrapped
the narrower operand in a `(safe-cast <joined> <own> ...)` annotation —
and the emitter peeled that wrapper silently, so the operands met as
u64 vs u32 and the generated crate failed `cargo build` with E0308.

Thirteen such errors in the digital-passport dogfood crate (surfaced
by add-digital-passport-dogfood-fixture's registration gate; every one
a typer-legal mixed-width site in assertCivilDateMatchesEpochDays /
the age predicate, including the `month >= 3 ? month - 3 : month + 9`
branch join whose arms ranged u32 vs u64).

The fix materialises the typer's widening at the boundaries where
widths must agree, instead of threading a joined type through the
renderer: comparison operands (`<= >= < > == !=`, both the expr-rust
clauses and the ctor-path coerce-cmp-operand-rust) and
conditional-expression branches render a wrapped narrower operand as
`((operand) as <wider>)` — a lossless zero-extension — while the wider
side keeps its own minimal-width rendering (`q * 4` stays u64). The
`+ - *` routes already emitted same-width operands via the mbits
result-width cast and needed no change. Bare integer literals still
peel silently (Rust infers their width), which keeps uniform-width
and literal-widening sites byte-identical: the full FIXTURES corpus
regenerates byte-identically outside two genuine mixed-width sites
(ternary-cond-fixture's `decrement` guard; the dogfood crate itself,
which also gains rustfmt normalisation of its staged-unformatted
lib.rs). The constructor body walker additionally skips the
declaration-only const statements the typer emits when a ctor const
RHS lifts its own temps — without that, `constructor { const diff =
base - q * 4; ... }` was wholly unwalkable.

Gates: examples/mixed_width_operand_fixture.compact (per-operator
mixed-width circuits, the exact dogfood shape, guarded subtraction,
impure + constructor routes) with executing parity tests driving
values past the 2^32 boundary so a truncating or wrongly-directed cast
cannot pass by merely compiling; cargo build -p tests-e2e-rust --tests
--locked green (zero E0308 in the dogfood crate); version 0.31.118.

+++ tests-e2e-rust/contracts/ternary-cond-fixture/lib.rs
@@ -314,7 +314,10 @@ pub mod pure_circuits {
 
     pub fn decrement(n: u64, flag: bool) -> Result<u64, CompactError> {
         let t = if flag { 1 } else { 0 };
-        compact_assert!((n >= t), "result of subtraction would be negative");
+        compact_assert!(
+            (n >= ((t) as u64)),
+            "result of subtraction would be negative"
+        );
         Ok(((n) as u64).wrapping_sub((t) as u64))
     }
